fastapp/requirements.txt
accelerate==0.34.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asn1crypto==1.5.1
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.13.0
coloredlogs==15.0.1
cryptography==43.0.3
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.1.0
fastcore==1.7.19
fastjsonschema==2.20.0
fastlite==0.0.13
fastsql==2.0.3
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.9.0
google-auth==2.35.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lazy_loader==0.4
llvmlite==0.43.0
Mako==1.3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.2
opencv-python-headless==4.10.0.84
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pg8000==1.31.2
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
psycopg==3.2.3
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMatting==1.1.12
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.8.0
python-multipart==0.0.16
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
replicate==1.0.2
requests==2.32.3
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
scramp==1.4.5
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlite-minutils==3.37.0.post4
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.3
tabulate==0.9.0
tifffile==2024.9.20
tinycss2==1.3.0
tokenizers==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.1
yarg==0.1.9
yarl==1.16.0
zipp==3.20.2
